% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_bg_rect.R
\name{base_bg_rect}
\alias{base_bg_rect}
\title{Calculates baseline window stats}
\usage{
base_bg_rect(dat, col)
}
\arguments{
\item{dat}{dataframe containing respirometry data}

\item{col}{column containing information on multiplexer chambers}
}
\value{
a dataframe with the beginning and end of each baseline sampling period
}
\description{
Determines the beginning and end of each time the multiplexer measures baseline data.
This is primarily a function for plotting and initial visial exploratory  analysis
of respirometry data.
}
\details{
This is a generic function:
}
\examples{
base_bg_rect(test_resp, "Marker")

}
