% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_corr.R
\name{baseline_corr}
\alias{baseline_corr}
\title{Baseline drift correction via spline interpolation}
\usage{
baseline_corr(dat, col1, col2, col3, conc1)
}
\arguments{
\item{dat}{dataframe containing respirometry data}

\item{col1}{column containing information on multiplexer chambers}

\item{col2}{column containing oxygen data}

\item{col3}{column containing sample number data}

\item{conc1}{a number specifying the known concetration or value for a channel. e.g. atmospheric O2 concentration is typically ~20.95%}
}
\value{
a dataframe containing drift correct oxygen and carbon dioxide data corrected using a
spline fit
}
\description{
When provided user defined baseline data, this function corrects for sensor drift i a specified channel
using three types of interpolation; linear, fmm spline, and natural spline. It is intended for the
user to compare multiple fits and determine which makes the most sense for drift correction.
}
\details{
This is a generic function:
}
\examples{
baseline_spl(test_resp, "Marker", "Oxygen","Sample")

}
