% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_spl.R
\name{baseline_spl}
\alias{baseline_spl}
\title{Calculates baseline window stats}
\usage{
baseline_spl(dat, col, col1, col2, col3)
}
\arguments{
\item{dat}{dataframe containing respirometry data}

\item{col}{column containing information on multiplexer chambers}

\item{col1}{column containing the oxygen data}

\item{col2}{column containing carbon dioxide data}

\item{col3}{column containing sample number data}
}
\value{
a dataframe with the beginning and end of each baseline sampling period
}
\description{
Determines the beginning and end of each time the multiplexer measures baseline data.
This is primarily a function for plotting and initial visial exploratory  analysis
of respirometry data.
}
\details{
This is a generic function:
}
\examples{
baseline_spl(test_resp, "Marker", "Oxygen", "CO2", "Sample")

}
