% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_plot.R
\name{drift_plot}
\alias{drift_plot}
\title{Baseline drift correction insepction plot}
\usage{
drift_plot(dat, dat2, val)
}
\arguments{
\item{dat}{dataframe containing spline fits of respirometry data, this is the output from the baseline_corr function}

\item{val}{the channel being drift correct, e.g. "Oxygen}

\item{dat1}{dataframe containing the baseline fit data, this is the output from the base_bg_rect function}
}
\value{
a dataframe containing drift correct oxygen and carbon dioxide data corrected using a
spline fit
}
\description{
This function provides a basic plot of the different spline and linear interpolation fits for the visual inspection of
the data.
}
\details{
This is a generic function:
}
\examples{
drift_plot (o2_corr, base_rect, "Oxygen")

}
