% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_9_7.R
\name{equation_9_7}
\alias{equation_9_7}
\title{Baseline drift correction inspetion plot}
\usage{
equation_9_7(dat, val1, val2)
}
\arguments{
\item{dat}{dataframe containing spline fits of respirometry data, this is the output from the baseline_corr function}

\item{dat1}{dataframe containing the baseline fit data, this is the output from the base_bg_rect function}

\item{val}{the channel being drift correct, e.g. "Oxygen"}
}
\value{
a dataframe containing drift correct oxygen and carbon dioxide data corrected using a
spline fit
}
\description{
This function provides a basic plot of the different spline and linear interpolation fits for the visual inspection of
the data.
}
\details{
This is a generic function:
}
\examples{
drift_plot (o2_corr, base_rect, "Oxygen")


}
