% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_multiplex.R
\name{fill_multiplex}
\alias{fill_multiplex}
\title{Replaces missing multiplexer channel values}
\usage{
fill_multiplex(dat, col, value, base)
}
\arguments{
\item{dat}{dataframe containing respirometry data}

\item{col}{column containing the information on multiplexer chambers}

\item{value}{The values in the original dataframe that are uninformative e.g.
(-999 or NA) that will be replaced with the actual sampled chamber value for each
row. This accepts single or multiple values (see examples below).}

\item{baseline}{the value for the baseline channel}
}
\value{
The original dataframe with formatted and corrected multiplexer channel IDs
}
\description{
Replaces missing values that are typically from Expedata multiplexer output with
the actual sample channel data.
}
\details{
This is a generic function:
}
\examples{
fill_multiplex(test_resp, Oxygen, -1, 66)
fill_multiple(test_resp, Oxygen, c(-1,-999), 66)

}
